pyproject.toml
[project]
name = "onboardly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bentoml>=1.4.5",
    "dlib==19.24.2",
    "easyocr>=1.7.2",
    "face-recognition>=1.3.0",
    "fastapi>=0.115.11",
    "just>=0.8.162",
    "locust>=2.33.2",
    "loguru>=0.7.3",
    "mediapipe>=0.10.21",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.10.6",
    "pymupdf>=1.25.3",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "uvicorn>=0.34.0",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cpu"},
]
torchvision = [
  { index = "pytorch-cpu" },
]


[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
